feat: 상세페이지 및 예약 페이지 기능 구현 렌더링 작업
