+++ src/components/my/MyReavervationCard.module.css
@@ -3,23 +3,23 @@
   display: flex;
   flex-wrap: wrap;
 }
-.contentsDiv {
+/* .contentsDiv {
   display: flex;
   flex-direction: column;
-  border: 1px solid var(--color-border-bg-light);
+
   width: 100%;
   height: 235px;
   padding: 10px;
   border-radius: 15px;
   margin: 0 10px 20px 0;
-}
+} */
 .roomList {
   display: flex;
 }
 .roomImageDiv {
   width: 80%;
   height: 170px;
-  margin-bottom: 5px;
+  margin: 0 5px 10px 5px;
 }
 .roomImage {
   width: 210px;
@@ -30,8 +30,11 @@
 }
 .roomContents {
   display: flex;
+  width: 30%;
+  min-width: 185px;
   flex-direction: column;
   gap: 15px;
+  margin-left: 20px;
 }
 .roomTitle {
   font-weight: 700;
