+++ src/components/reservation/ImgageDiv.jsx
@@ -2,21 +2,12 @@ import React, { useState } from "react";
 
 import classes from "./ImgageDiv.module.css";
 
-const ImgageDiv = () => {
+const ImgageDiv = ({ imageArr }) => {
   return (
     <div style={{ margin: "25px 0" }}>
       <div className={classes.imageGroup}>
-        <div className={classes.image1}></div>
-        <div className={classes.image2}>
-          <div className={classes.image2_1}>
-            <div className={classes.image2_1_1}></div>
-            <div className={classes.image2_1_1}></div>
-          </div>
-          <div className={classes.image2_2}>
-            <div className={classes.image2_2_1}></div>
-            <div className={classes.image2_2_2}></div>
-          </div>
-        </div>
+        <img src={imageArr[0]} className={classes.image1} />
+        <img src={imageArr[1]} className={classes.image2} />
       </div>
     </div>
   );
