+++ src/components/reservation/ImgageDiv.module.css
@@ -1,52 +1,14 @@
 .imageGroup {
   display: flex;
+  gap: 20px;
 }
 .image1 {
   width: 50%;
   height: 370px;
-  background-image: url("../../img/testImg/testImg1.png");
-  background-size: cover;
-  border-radius: 15px 0 0 0;
+  border-radius: 15px 0 0 15px;
 }
 .image2 {
-  display: flex;
   width: 50%;
-}
-.image2_1 {
-  display: flex;
-  width: 50%;
-  flex-direction: column;
-  margin-left: 10px;
-}
-.image2_1_1 {
-  width: 100%;
-  height: 180px;
-  background-image: url("../../img/testImg/testImg1.png");
-  background-size: cover;
-  margin-bottom: 10px;
-}
-.image2_1_2 {
-  width: 100%;
-  height: 180px;
-  background-image: url("../../img/testImg/testImg3.png");
-  background-size: cover;
-}
-.image2_2 {
-  width: 50%;
-  margin-left: 10px;
-}
-.image2_2_1 {
-  width: 100%;
-  height: 180px;
-  background-image: url("../../img/testImg/testImg4.png");
-  background-size: cover;
-  border-radius: 0 15px 0 0;
-  margin-bottom: 10px;
-}
-.image2_2_2 {
-  width: 100%;
-  height: 180px;
-  background-image: url("../../img/testImg/testImg5.png");
-  background-size: cover;
-  border-radius: 0 0 15px 0;
+  height: 370px;
+  border-radius: 0 15px 15px 0;
 }
